🚚 Migrate to TypeScript
Add TypeScript, fix TypeScript errors and upgrade dependencies
