+++ src/components/UI/Footer.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Footer = () => {
+  return (
+    <div className="relative bottom-0 w-full m-0">
+      <footer className="flex justify-center items-center bg-black h-14">
+        <p className="font-medium text-white text-center ">
+          Copyright &copy; 2024 Glaiza Cano
+        </p>
+      </footer>
+    </div>
+  );
+};
+
+export default Footer;
