--- src/containers/Projects/TechnologyStack.js
@@ -1,62 +1,0 @@
-import React from 'react';
-
-const items = [
-  {
-    name: 'CSS',
-    icon: '',
-  },
-  {
-    name: 'C#',
-    icon: '',
-  },
-  {
-    name: 'HTML',
-    icon: '',
-  },
-  {
-    name: 'Invision Community',
-    icon: '',
-  },
-  {
-    name: 'JavaScript',
-    icon: '',
-  },
-  {
-    name: 'jQuery',
-    icon: '',
-  },
-  {
-    name: 'Netlify',
-    icon: '',
-  },
-  {
-    name: 'Node.js',
-    icon: '',
-  },
-  {
-    name: 'PHP',
-    icon: '',
-  },
-  {
-    name: 'React.js',
-    icon: '',
-  },
-  {
-    name: 'Tailwind CSS',
-    icon: '',
-  },
-  {
-    name: 'WinForms',
-    icon: '',
-  },
-  {
-    name: 'Wordpress',
-    icon: '',
-  },
-];
-
-const TechnologyStack = () => {
-  return <div>Technology stack:</div>;
-};
-
-export default TechnologyStack;
