--- src/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './styles/index.css';
-import App from './containers/App';
-
-const root = document.getElementById('root');
-
-ReactDOM.render(<App />, root);
